fix: get20Rankings 수정

+++ src/main/java/swyp_8th/bungmakase_backend/dto/profile/RankingResponseDto.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class RankingResponseDto {
-    private int rank; // user ranking start from 1
+    private Long rank; // user ranking start from 1
     private String nickname; // user nickname
     private Long level; // user level
     private Long recentBungCount; // total fish-bun that user ate
